Modularize helper method
