+++ app/helpers/application_helper.rb
@@ -1,12 +1,4 @@
 module ApplicationHelper
-  def divide_by_comma(users)
-    usernames = []
-    users.each do |u|
-      usernames << u.name
-    end
-    usernames.join(", ")
-  end
-
   def display_latest_message_or_status_message(group)
     if group.messages.any?
       if group.messages.last.body.empty?
