+++ app/helpers/users_helper.rb
@@ -1,0 +1,9 @@
+module UsersHelper
+  def divide_by_comma(users)
+    usernames = []
+    users.each do |u|
+      usernames << u.name
+    end
+    usernames.join(", ")
+  end
+end
